revert change, add test
