+++ patches/0038-Non-idempotent-sns-topic-creation.patch
