+++ provider/provider_test.go
@@ -3,12 +3,16 @@ package provider
 
 import (
 	"context"
+	"math/rand"
 	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 
 	"github.com/pulumi/providertest"
+	"github.com/pulumi/providertest/pulumitest"
+	"github.com/pulumi/providertest/pulumitest/opttest"
 	pfbridge "github.com/pulumi/pulumi-terraform-bridge/pf/tfbridge"
 	pulumirpc "github.com/pulumi/pulumi/sdk/v3/proto/go"
 
@@ -55,4 +59,30 @@ func getEnvRegion(t *testing.T) string {
 		envRegion = "us-west-2"
 	}
 	return envRegion
+}
+
+func pulumiTest(t *testing.T, dir string) *pulumitest.PulumiTest {
+	if testing.Short() {
+		t.Skipf("Skipping in testing.Short() mode, assuming this is a CI run without AWS creds")
+		return nil
+	}
+	cwd, err := os.Getwd()
+	if err != nil {
+		t.Error(err)
+	}
+	ptest := pulumitest.NewPulumiTest(t, dir,
+		opttest.LocalProviderPath("aws", filepath.Join(cwd, "..", "bin")),
+	)
+
+	return ptest
+}
+
+var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
+
+func randSeq(n int) string {
+	b := make([]rune, n)
+	for i := range b {
+		b[i] = letters[rand.Intn(len(letters))]
+	}
+	return string(b)
 }
