+++ provider/provider_yaml_test.go
@@ -6,14 +6,10 @@
 package provider
 
 import (
-	"math/rand"
-	"os"
 	"path/filepath"
 	"testing"
 
 	"github.com/pulumi/providertest"
-	"github.com/pulumi/providertest/pulumitest"
-	"github.com/pulumi/providertest/pulumitest/opttest"
 	"github.com/stretchr/testify/require"
 )
 
@@ -88,32 +84,6 @@ func TestKmsKey(t *testing.T) {
 
 func TestSecretsManagerSecret(t *testing.T) {
 	test(t, filepath.Join("test-programs", "secretsmanager-secret"))
-}
-
-var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
-
-func randSeq(n int) string {
-	b := make([]rune, n)
-	for i := range b {
-		b[i] = letters[rand.Intn(len(letters))]
-	}
-	return string(b)
-}
-
-func pulumiTest(t *testing.T, dir string) *pulumitest.PulumiTest {
-	if testing.Short() {
-		t.Skipf("Skipping in testing.Short() mode, assuming this is a CI run without AWS creds")
-		return nil
-	}
-	cwd, err := os.Getwd()
-	if err != nil {
-		t.Error(err)
-	}
-	ptest := pulumitest.NewPulumiTest(t, dir,
-		opttest.LocalProviderPath("aws", filepath.Join(cwd, "..", "bin")),
-	)
-
-	return ptest
 }
 
 func TestNonIdempotentSnsTopic(t *testing.T) {
