+++ provider/test-programs/sns-topic-replacement/Pulumi.yaml
@@ -1,0 +1,3 @@
+name: sns-topic-replacement
+runtime: go
+description: A minimal AWS Go Pulumi program
